+++ tests/api/CalendarApiCest.php
@@ -4,15 +4,36 @@ namespace Foodsharing\api;
 
 use ApiTester;
 use Codeception\Util\HttpCode;
+use Foodsharing\Modules\Event\InvitationStatus;
 
 class CalendarApiCest
 {
     private const TEST_TOKEN = '1234567890';
     private $user;
+    private $user2;
+    private $region;
+    private array $acceptedEvent;
+    private array $invitedEvent;
 
     public function _before(ApiTester $I)
     {
+        $this->region = $I->createRegion();
+
         $this->user = $I->createFoodsharer();
+        $I->addRegionMember($this->region['id'], $this->user['id']);
+
+        $this->user2 = $I->createFoodsharer();
+        $I->addRegionMember($this->region['id'], $this->user2['id']);
+
+        $this->acceptedEvent = $I->createEvents($this->region['id'], $this->user2['id']);
+        $I->addEventInvitation($this->acceptedEvent['id'], $this->user['id'], [
+            'status' => InvitationStatus::ACCEPTED
+        ]);
+
+        $this->invitedEvent = $I->createEvents($this->region['id'], $this->user2['id']);
+        $I->addEventInvitation($this->invitedEvent['id'], $this->user['id'], [
+            'status' => InvitationStatus::INVITED
+        ]);
     }
 
     public function canNotAccessApiWithoutLogin(ApiTester $I)
@@ -107,5 +128,26 @@ class CalendarApiCest
         $I->sendGet('api/calendar/' . self::TEST_TOKEN);
         $I->seeResponseCodeIs(HttpCode::OK);
         $I->seeResponseContains('BEGIN:VCALENDAR');
+        $I->seeResponseContains(substr($this->invitedEvent['name'], 0, 10));
+        $I->seeResponseContains(substr($this->acceptedEvent['name'], 0, 10));
+    }
+
+    public function canFilterOutInvitations(ApiTester $I)
+    {
+        $I->haveInDatabase('fs_apitoken', [
+            'foodsaver_id' => $this->user['id'],
+            'token' => self::TEST_TOKEN
+        ]);
+
+        $I->login($this->user['email']);
+        $I->sendGet('api/calendar/' . self::TEST_TOKEN);
+        $I->seeResponseCodeIs(HttpCode::OK);
+        $I->seeResponseContains(substr($this->acceptedEvent['name'], 0, 10));
+        $I->seeResponseContains(substr($this->invitedEvent['name'], 0, 10));
+
+        $I->sendGet('api/calendar/' . self::TEST_TOKEN . '?events=answered');
+        $I->seeResponseCodeIs(HttpCode::OK);
+        $I->seeResponseContains(substr($this->acceptedEvent['name'], 0, 10));
+        $I->cantSeeResponseContains(substr($this->invitedEvent['name'], 0, 10));
     }
 }